+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 191:
 * Fix buffers_cat uninitialized warning
 * Fix static_string uninitialized warning
 * Fix warning in is_ssl_handshake
+* example/cppcon2018 only requires C++11
 
 --------------------------------------------------------------------------------